Moved new cards to the cards folder and added habitat types to cards
