+++ G9_Earth_Base/src/crm/cards/CardAmericanSycamore.java
@@ -1,7 +1,6 @@
 package crm.cards;
 
-import crm.CardType;
-import crm.twoActionCard;
+import crm.*;
 
 public class CardAmericanSycamore extends twoActionCard {
