+++ G9_Earth_Base/src/crm/cards/CardShrubland.java
@@ -1,4 +1,6 @@
-package crm;
+package crm.cards;
+
+import crm.*;
 
 public class CardShrubland extends twoActionCard {
