+++ G9_Earth_Base/src/crm/cards/FallenSequoia.java
@@ -7,6 +7,7 @@ public class FallenSequoia extends oneActionCard{
     {
         setName("Fallen Sequoia");
         setType(CardType.TERRAIN);
+        setHabitatType(false, true, true, true);
         setSoilCost(2);
         setBasePointValue(9);
         setActionDesc("All Mushroom cost 1 less soil (cannot cost less than 0)");
